Move TestRegressor to its own file
